WIP: Move more codes
