+++ Cargo.toml
@@ -44,6 +44,7 @@ swc_ecma_transforms = {version = "0.46.0", path = "./ecmascript/transforms", fea
 ]}
 swc_ecma_utils = {version = "0.34.1", path = "./ecmascript/utils"}
 swc_ecma_visit = {version = "0.29.1", path = "./ecmascript/visit"}
+swc_node_base = { path = "./node/base" }
 swc_visit = {version = "0.2.3", path = "./visit"}
 
 [dev-dependencies]
