+++ node/base/Cargo.toml
@@ -1,13 +1,18 @@
 [package]
 authors = ["강동윤 <kdy1997.dev@gmail.com>"]
+description = "Speedy web compiler"
+documentation = "https://rustdoc.swc.rs/swc_node_base/"
 edition = "2018"
+license = "Apache-2.0/MIT"
 name = "swc_node_base"
-publish = false
+repository = "https://github.com/swc-project/swc.git"
 version = "0.1.0"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+dashmap = "4.0.2"
+swc_common = {version = "0.10", path = "../../common"}
 
 [target.'cfg(all(unix, not(target_env = "musl"), not(target_arch = "aarch64")))'.dependencies]
 jemallocator = {version = "0.3", features = ["disable_initial_exec_tls"]}
