+++ node/base/src/lib.rs
@@ -2,7 +2,6 @@
 //!
 //! The swc crates related to the  node binding should depend on this crate.
 
-
 #[cfg(all(unix, not(target_env = "musl"), not(target_arch = "aarch64")))]
 #[global_allocator]
 static ALLOC: jemallocator::Jemalloc = jemallocator::Jemalloc;
@@ -10,3 +9,5 @@ static ALLOC: jemallocator::Jemalloc = jemallocator::Jemalloc;
 #[cfg(windows)]
 #[global_allocator]
 static ALLOC: mimalloc::MiMalloc = mimalloc::MiMalloc;
+
+pub mod comments;
